Modificaciones a plan de adquisiciones y control de adquisiciones
