+++ sigpro/WebContent/app/components/main/main.controller.js
@@ -178,8 +178,8 @@ app.config(['$routeProvider', '$locationProvider','FlashProvider',function ($rou
             .when('/responsablerol/:reiniciar_vista?',{
             	template: '<div load-on-demand="\'responsablerolController\'" class="all_page"></div>'
             })
-            .when('/planadquisiciones/',{
-            	template: '<div load-on-demand="\'planAdquisicionesController\'" class="all_page"></div>'
+            .when('/controladquisiciones/',{
+            	template: '<div load-on-demand="\'controlAdquisicionesController\'" class="all_page"></div>'
             })
             .when('/planejecucion/',{
             	template: '<div load-on-demand="\'planejecucionController\'" class="all_page"></div>'
@@ -196,8 +196,8 @@ app.config(['$routeProvider', '$locationProvider','FlashProvider',function ($rou
             .when('/tipoadquisicion/:reiniciar_vista?',{
             	template: '<div load-on-demand="\'tipoAdquisicionController\'" class="all_page"></div>'
             })
-            .when('/adquisicion/:objeto_id/:objeto_tipo/:reiniciar_vista?',{
-            	template: '<div load-on-demand="\'adquisicionController\'" class="all_page"></div>'
+            .when('/planadquisiciones/',{
+            	template: '<div load-on-demand="\'planAdquisicionesController\'" class="all_page"></div>'
             })
             .when("/:redireccion?",{
             	controller:"MainController"
@@ -475,9 +475,9 @@ app.config(['$loadOnDemandProvider', function ($loadOnDemandProvider) {
 	    	   script: '/app/components/responsablerol/responsablerol.controller.js',
 	    	   template: '/app/components/responsablerol/responsablerol.jsp'
 	       },{
-	    	   name: 'planAdquisicionesController',
-	    	   script: '/app/components/reportes/planadquisiciones/planadquisiciones.controller.js',
-	    	   template: '/app/components/reportes/planadquisiciones/planadquisiciones.jsp'
+	    	   name: 'controlAdquisicionesController',
+	    	   script: '/app/components/reportes/controladquisiciones/controladquisiciones.controller.js',
+	    	   template: '/app/components/reportes/controladquisiciones/controladquisiciones.jsp'
 	       },{
 	    	   name: 'planejecucionController',
 	    	   script: '/app/components/reportes/planejecucion/planejecucion.controller.js',
@@ -508,9 +508,9 @@ app.config(['$loadOnDemandProvider', function ($loadOnDemandProvider) {
 	    	   template: '/app/components/tipoadquisicion/tipoAdquisicion.jsp'
 	       },
 	       {
-	    	   name: 'adquisicionController',
-	    	   script: '/app/components/adquisicion/adquisicion.controller.js',
-	    	   template: '/app/components/adquisicion/adquisicion.jsp'
+	    	   name: 'planAdquisicionesController',
+	    	   script: '/app/components/reportes/planadquisiciones/planAdquisiciones.controller.js',
+	    	   template: '/app/components/reportes/planadquisiciones/planAdquisiciones.jsp'
 	       }
 
 	   ];
