+++ sigpro/WebContent/app/components/reportes/controladquisiciones/controladquisiciones.controller.js
@@ -1,5 +1,5 @@
-var app = angular.module('planAdquisicionesController', ['ngTouch','ngAnimate','ui.grid.edit', 'ui.grid.rowEdit','ui.utils.masks']);
-app.controller('planAdquisicionesController',['$scope', '$http', '$interval', 'uiGridTreeViewConstants','Utilidades','i18nService','uiGridConstants','$timeout', 'uiGridTreeBaseService', '$q','dialogoConfirmacion', '$filter','$uibModal',
+var app = angular.module('controlAdquisicionesController', ['ngTouch','ngAnimate','ui.grid.edit', 'ui.grid.rowEdit','ui.utils.masks']);
+app.controller('controlAdquisicionesController',['$scope', '$http', '$interval', 'uiGridTreeViewConstants','Utilidades','i18nService','uiGridConstants','$timeout', 'uiGridTreeBaseService', '$q','dialogoConfirmacion', '$filter','$uibModal',
 	function($scope, $http, $interval, uiGridTreeViewConstants,$utilidades,i18nService,uiGridConstants,$timeout, uiGridTreeBaseService, $q, $dialogoConfirmacion, $filter,$uibModal) {
 	var mi = this;
 	var anioFiscal = new Date();
@@ -621,7 +621,7 @@ app.controller('planAdquisicionesController',['$scope', '$http', '$interval', 'u
 			estructuraGuardar += "|";
 		}
 		
-		$http.post('/SPlanAdquisiciones', {
+		$http.post('/SControlAdquisiciones', {
 			accion: 'guardarPlan',
 			data : estructuraGuardar,
 			t:moment().unix()
@@ -672,7 +672,7 @@ app.controller('planAdquisicionesController',['$scope', '$http', '$interval', 'u
 			mi.mostrarCargando = true;
 			mi.mostrarTablas = false;
 			mi.idPrestamo = mi.prestamo.value;
-			$http.post('/SPlanAdquisiciones',{
+			$http.post('/SControlAdquisiciones',{
 				accion: 'generarPlan',
 				idPrestamo: mi.idPrestamo
 			}).success(function(response){
@@ -732,7 +732,7 @@ app.controller('planAdquisicionesController',['$scope', '$http', '$interval', 'u
 	}
 	
 	mi.exportarExcel = function(){
-		$http.post('/SPlanAdquisiciones', { 
+		$http.post('/SControlAdquisiciones', { 
 			 accion: 'exportarExcel', 
 			 idPrestamo: mi.idPrestamo,
 			 informeCompleto: mi.informeCompleto,			 
@@ -751,7 +751,7 @@ app.controller('planAdquisicionesController',['$scope', '$http', '$interval', 'u
 		};
 	
 	mi.exportarPdf=function(){
-		$http.post('/SPlanAdquisiciones', { 
+		$http.post('/SControlAdquisiciones', { 
 			 accion: 'exportarPdf', 
 			 idPrestamo: mi.idPrestamo,
 			 informeCompleto: mi.informeCompleto,			 
@@ -917,7 +917,7 @@ function ($uibModalInstance, $scope, $http, $interval,
 						$uibModalInstance.close({success: false, numeroContrato: mi.numeroContrato, montoContrato: mi.montoContrato});
 			});
 			
-			$http.post('/SPlanAdquisiciones', 
+			$http.post('/SControlAdquisiciones', 
 				{
 					accion: 'guardarMontoContrato', 
 					numeroContrato: mi.numeroContrato,
